mejoras en la resitriccion a paginas sin logeo previo TERMINADO

+++ src/main/webapp/lista-usuarios.jsp
@@ -50,7 +50,7 @@
 								
 					</td>
 					<td>
-						<a href="eliminar-usu?id=${u.id}">
+						<a onclick="confirm('¿ Deseas borrar el usuario ${u.nombre}?')" href="eliminar-usu?id=${u.id}">
 							<i class="fas fa-trash text-danger"></i>
 						</a>
 					</td>
